+++ components/site/src/tpls.rs
@@ -15,8 +15,10 @@ pub fn register_early_global_fns(site: &mut Site) {
             site.config.clone(),
             site.permalinks.clone(),
             site.output_path.clone(),
+            site.library.colocated_assets.clone(),
         ),
     );
+
     site.tera.register_function(
         "resize_image",
         functions::ResizeImage::new(
@@ -58,6 +60,7 @@ pub fn register_early_global_fns(site: &mut Site) {
         filters::MarkdownFilter::new(
             site.config.clone(),
             site.permalinks.clone(),
+            site.library.colocated_assets.clone(),
             site.tera.clone(),
         ),
     );